refactor: reorganize utility functions; consolidate bucket and seller profile utilities; remove unused files and update imports
