+++ web-client/src/utils/bookings.ts
@@ -1,28 +1,24 @@
 import { format, addMinutes } from "date-fns";
 import supabase from "./supabase";
-import { getUser } from "./auth";
 
 
 export async function bookSlot(
+  userId: string,
   sellerId: string,
   serviceId: number,
   date: Date,
   slot: string,
 ): Promise<{ success: boolean; message: string }> {
   try {
-    // Get the current user
-    const user = await getUser();
 
-    if (!user) {
+    if (!userId) {
       return { success: false, message: "User not authenticated" };
     }
-
-    const userId = user.id;
 
     // Fetch the selected service (name + duration)
     const { data: service, error: serviceError } = await supabase
       .from("service")
-      .select("name, duration")
+      .select("name,duration")
       .eq("id", serviceId)
       .single();
 
@@ -33,10 +29,13 @@ export async function bookSlot(
     const serviceName = service.name;
     const duration = service.duration;
 
+    
+
     // Calculate start and end time
     const dateString = format(date, "yyyy-MM-dd");
     const startDateTime = new Date(`${dateString}T${slot}`);
     const endDateTime = addMinutes(startDateTime, duration);
+
 
     // Insert booking
     const { error: insertError } = await supabase.from("bookings").insert([
@@ -44,7 +43,6 @@ export async function bookSlot(
         seller_id: sellerId,
         user_id: userId,
         service_id: serviceId,
-        service_name: serviceName,
         start_time: startDateTime.toISOString(),
         end_time: endDateTime.toISOString(),
       },
@@ -64,4 +62,17 @@ export async function bookSlot(
       message: err.message || "Unexpected error occurred",
     };
   }
+}
+
+export async function cancelBooking(bookingId: string) {
+  const { error } = await supabase
+    .from("bookings")
+    .update({ status: "cancelled" })
+    .eq("id", bookingId);
+
+  if (error) {
+    return { success: false, message: error.message };
+  }
+
+  return { success: true, message: "Booking cancelled" };
 }
