+++ web-client/src/utils/bucket.ts
