+++ web-client/src/utils/seller.ts
@@ -1,4 +1,4 @@
-import { Seller } from "@/types/types";
+import { Seller, Service } from "@/types/types";
 import supabase from "./supabase";
 
 export async function getSellers(): Promise<Seller[]> {
@@ -31,3 +31,26 @@ export async function getAvailableDates(sellerId: string): Promise<string[]> {
 
   return data.map((row: { day: string }) => row.day.split('T')[0]);
 }
+
+export const fetchServices = async (
+  userId: string,
+): Promise<{ data: Service[]; error: string | null }> => {
+  try {
+    const { data, error } = await supabase
+      .from("service")
+      .select("id, name, description, price, category, duration, user_id")
+      .eq("user_id", userId);
+
+    if (error) {
+      throw new Error("Failed to fetch services. Please try again later.");
+    }
+
+    return { data: data || [], error: null };
+  } catch (err) {
+    return {
+      data: [],
+      error:
+        err instanceof Error ? err.message : "An unexpected error occurred.",
+    };
+  }
+};
